Added workaround for SleekXMPP reconnect 'bug'
If a reconnect is attempted after access token is expired, SleekXMPP's
SASL auth fails and the client is stopped. (Note Google's access tokens
expire after one hour, so this is pretty much guaranteed to happen as
long as the internet connection is not flaky) Get around this by
updating the access token whenever the client establishes a connection
to the XMPP server. By product is that access token is requested twice
upon startup.